refactor: theme 응집
