+++ src/App.tsx
@@ -1,11 +1,6 @@
 import React, { useState } from "react";
 import { generateItems, renderLog } from "./utils";
-import {
-  User,
-  UserContext,
-  UserContextType,
-  useUserContext,
-} from "./context/user";
+import { useUserContext } from "./context/user";
 import { useNotificationContext } from "./context/notification/hook";
 import { Providers } from "./context/providers";
 import { useThemeContext } from "./context/theme";
@@ -244,7 +239,6 @@ export const NotificationSystem: React.FC = () => {
 // 메인 App 컴포넌트
 const App: React.FC = () => {
   const [items, setItems] = useState(generateItems(1000));
-  const [user, setUser] = useState<User | null>(null);
 
   const addItems = () => {
     setItems((prevItems) => [
@@ -253,42 +247,24 @@ const App: React.FC = () => {
     ]);
   };
 
-  const login = (email: string) => {
-    setUser({ id: 1, name: "홍길동", email });
-    // addNotification("성공적으로 로그인되었습니다", "success");
-  };
-
-  const logout = () => {
-    setUser(null);
-    // addNotification("로그아웃되었습니다", "info");
-  };
-
-  const userContextValue: UserContextType = {
-    user,
-    login,
-    logout,
-  };
-
   return (
     <Providers>
-      <UserContext.Provider value={userContextValue}>
-        <div
-        // className={`min-h-screen ${theme === "light" ? "bg-gray-100" : "bg-gray-900 text-white"}`}
-        >
-          <Header />
-          <div className="container mx-auto px-4 py-8">
-            <div className="flex flex-col md:flex-row">
-              <div className="w-full md:w-1/2 md:pr-4">
-                <ItemList items={items} onAddItemsClick={addItems} />
-              </div>
-              <div className="w-full md:w-1/2 md:pl-4">
-                <ComplexForm />
-              </div>
+      <div
+      // className={`min-h-screen ${theme === "light" ? "bg-gray-100" : "bg-gray-900 text-white"}`}
+      >
+        <Header />
+        <div className="container mx-auto px-4 py-8">
+          <div className="flex flex-col md:flex-row">
+            <div className="w-full md:w-1/2 md:pr-4">
+              <ItemList items={items} onAddItemsClick={addItems} />
+            </div>
+            <div className="w-full md:w-1/2 md:pl-4">
+              <ComplexForm />
             </div>
           </div>
-          <NotificationSystem />
         </div>
-      </UserContext.Provider>
+        <NotificationSystem />
+      </div>
     </Providers>
   );
 };
