+++ src/context/providers.tsx
@@ -1,10 +1,13 @@
 import { NotificationProvider } from "./notification/provider";
 import { ThemeProvider } from "./theme/provider";
+import { UserProvider } from "./user/providers";
 
 export const Providers = ({ children }: { children: React.ReactNode }) => {
   return (
     <NotificationProvider>
-      <ThemeProvider>{children}</ThemeProvider>
+      <UserProvider>
+        <ThemeProvider>{children}</ThemeProvider>
+      </UserProvider>
     </NotificationProvider>
   );
 };
