+++ src/context/user/providers.tsx
@@ -1,0 +1,22 @@
+import { useState } from "react";
+import { User, UserContext } from ".";
+
+export const UserProvider = ({ children }: { children: React.ReactNode }) => {
+  const [user, setUser] = useState<User | null>(null);
+
+  const login = (email: string) => {
+    setUser({ id: 1, name: "홍길동", email });
+    // addNotification("성공적으로 로그인되었습니다", "success");
+  };
+
+  const logout = () => {
+    setUser(null);
+    // addNotification("로그아웃되었습니다", "info");
+  };
+
+  return (
+    <UserContext.Provider value={{ user, login, logout }}>
+      {children}
+    </UserContext.Provider>
+  );
+};
